pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "tiny-alignment-studio"
version = "0.1.0"
description = "Educational RLHF framework for learning alignment techniques"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "torch>=2.1.0",
    "transformers>=4.36.0",
    "trl>=0.7.0",
    "peft>=0.7.0",
    "bitsandbytes>=0.41.0",
    "datasets>=2.16.0",
    "accelerate>=0.25.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0",
    "streamlit>=1.30.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
]
wandb = [
    "wandb>=0.16.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "N",
    "UP",
    "B",
    "SIM",
    "TCH",
    "RUF",
]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["src"]
